requirements.txt
-e .
aws-cdk.aws-iam~=1.130.0
aws-cdk.aws-lambda~=1.130.0
aws_cdk.aws_logs~=1.130.0